+++ ngClient/App/controllers/content/closingNotesController.js
@@ -3,9 +3,9 @@
 
     var controllerId = 'ClosingNotesController';
     angular.module('main')
-        .controller(controllerId, ['userFactory', '$scope', '$sce', '$location', 'logger', ClosingNotesController]);
+        .controller(controllerId, ['userFactory', '$scope', '$sce', '$location', 'disqusShortname', 'logger', ClosingNotesController]);
 
-    function ClosingNotesController(userFactory, $scope, $sce, $location, logger) {
+    function ClosingNotesController(userFactory, $scope, $sce, $location, disqusShortname, logger) {
 
         // Logger
         logger = logger.forSource(controllerId);
@@ -16,6 +16,10 @@
         vm.isLocalhost = $location.$$host === 'localhost';
         vm.displayBankTransfer = false;
         vm.toggleBankTransfer = toggleBankTransfer;
+
+        vm.disqusShortname = disqusShortname;
+        vm.disqusId = 'wealth_economy_11'; // ?
+        vm.disqusUrl = $location.url(); // ?
 
         _init();