+++ ngClient/App/main.js
@@ -12,6 +12,7 @@
         'ngSanitize',
         'breeze.angular',
         'angular-google-analytics',
+        'angularUtils.directives.dirDisqus',
         'ui.bootstrap',
         'highcharts-ng'
     ]);