+++ SolutionItems/Documents/Release Notes.md
@@ -3,8 +3,8 @@
 **0.43**
 
 * gulp: auto jshint check, bundling, minifying for js & css files
-* "angular-google-analytics" module for handling google analytics
-* disqus was moved under angular
+* "angular-google-analytics" for google analytics
+* "angular-utils-disqus" for disqus
 * jshint fixes
 * ResourcePoolEdit(Controller) was renamed to ResourcePoolManage(Controller) to fix the conflict with generated one